+++ libs/jacdac-drivers/accelerometerclient.ts
@@ -68,31 +68,6 @@ namespace jacdac {
         onEvent(gesture: JDGesture, handler: () => void) {
             this.registerEvent(gesture, handler);
         }
-
-        static debugView(): DebugView {
-            return new AccelerometerDebugView();
-        }
-    }
-
-    class AccelerometerDebugView extends SensorDebugView {
-        constructor() {
-            super("acc", jacdac.ACCELEROMETER_DEVICE_CLASS);
-        }
-
-        renderEvent(data: number): string {
-            switch(data) {
-                case JDGesture.Shake: return "shake";
-                case JDGesture.FreeFall: return "freefall";
-            }
-            return "";
-        }
-
-        renderState(data: Buffer): string {
-            const x = data.getNumber(NumberFormat.UInt16LE, 0);
-            const y = data.getNumber(NumberFormat.UInt16LE, 2);
-            const z = data.getNumber(NumberFormat.UInt16LE, 4);
-            return `${x} ${y} ${z}`;
-        }
     }
 
     //% fixedInstance whenUsed block="accelerometer"
