+++ libs/jacdac/pxt.json
@@ -17,6 +17,7 @@
         "console.ts",
         "bridge.ts",
         "debugger.ts",
+        "debugviews.ts",
         "ns.ts"
     ],
     "testFiles": [
